requirements/test.txt
gitpython==3.1.44
packaging==24.2
pytest==8.3.4
pytest-xdist==3.6.1
pytest-cov==6.0.0
ruff==0.9.1